### 2024-11-12 #7460 @editorialbot
[PRE REVIEW]: Slicer-Liver: A 3D Slicer Extension for Liver Surgery Planning
**Submitting author:** @RuoyanMeng (<a href="http://orcid.org/0000-0003-0246-0011">Ruoyan Meng</a>)
**Repository:** https://github.com/ALive-research/Slicer-Liver
**Branch with paper.md** (empty if default branch):
**Version:** v1.0
