### 2025-01-14 2e05bdc
DOC: add contributing.md and fix missing doi
