+++ Paper/bibliography.bib
@@ -59,8 +59,11 @@
   title={Visual computing for medicine: theory, algorithms, and applications},
   author={Preim, Bernhard and Botha, Charl P},
   year={2013},
-  publisher={Newnes}
+  publisher={Newnes},
+  isbn = {978-0-12-415873-3},
+  doi = {10.1016/C2011-0-05785-X}
 }
+
 @article{Warmann:2016,
   title={Computer-assisted surgery planning in children with complex liver tumors identifies variability of the classical Couinaud classification},
   author={Warmann, Steven W and Schenk, Andrea and Schaefer, Juergen F and Ebinger, Martin and Blumenstock, Gunnar and Tsiflikas, Ilias and Fuchs, Joerg},
@@ -69,6 +72,7 @@
   number={11},
   pages={1801--1806},
   year={2016},
+  doi = {10.1016/j.jpedsurg.2016.05.018},
   publisher={Elsevier}
 }
 @article{Bismuth:2013,
@@ -79,6 +83,7 @@
   number={3},
   pages={383--386},
   year={2013},
+  doi={10.1097/SLA.0b013e31827f171f},
   publisher={LWW}
 }
 @incollection{Kikinis:2013,
@@ -87,6 +92,7 @@
   booktitle={Intraoperative imaging and image-guided therapy},
   pages={277--289},
   year={2013},
+  doi={10.1007/978-1-4614-7657-3_19},
   publisher={Springer}
 }
 
@@ -127,7 +133,8 @@ number = {1},
 pages = {17-48},
 keywords = {cancer cases, cancer statistics, death rates, incidence, mortality},
 abstract = {Abstract Each year, the American Cancer Society estimates the numbers of new cancer cases and deaths in the United States and compiles the most recent data on population-based cancer occurrence and outcomes using incidence data collected by central cancer registries and mortality data collected by the National Center for Health Statistics. In 2023, 1,958,310 new cancer cases and 609,820 cancer deaths are projected to occur in the United States. Cancer incidence increased for prostate cancer by 3\% annually from 2014 through 2019 after two decades of decline, translating to an additional 99,000 new cases; otherwise, however, incidence trends were more favorable in men compared to women. For example, lung cancer in women decreased at one half the pace of men (1.1\% vs. 2.6\% annually) from 2015 through 2019, and breast and uterine corpus cancers continued to increase, as did liver cancer and melanoma, both of which stabilized in men aged 50 years and older and declined in younger men. However, a 65\% drop in cervical cancer incidence during 2012 through 2019 among women in their early 20s, the first cohort to receive the human papillomavirus vaccine, foreshadows steep reductions in the burden of human papillomavirus-associated cancers, the majority of which occur in women. Despite the pandemic, and in contrast with other leading causes of death, the cancer death rate continued to decline from 2019 to 2020 (by 1.5\%), contributing to a 33\% overall reduction since 1991 and an estimated 3.8 million deaths averted. This progress increasingly reflects advances in treatment, which are particularly evident in the rapid declines in mortality (approximately 2\% annually during 2016 through 2020) for leukemia, melanoma, and kidney cancer, despite stable/increasing incidence, and accelerated declines for lung cancer. In summary, although cancer mortality rates continue to decline, future progress may be attenuated by rising incidence for breast, prostate, and uterine corpus cancers, which also happen to have the largest racial disparities in mortality.},
-year = {2023}
+year = {2023},
+doi = {10.3322/caac.21763}
 }
 
 @article{Petrowsky2020,
