### 2025-01-28 #7460 @Kevin-Mattheus-Moerman
Thanks @RuoyanMeng for working on those points. Another minor point, I recommend that you edit your license text to exactly match the standard license text (see also: https://opensource.org/license/BSD-3-clause) this way our system and also GitHub will recognise it properly.
