### 2025-02-05 #7460 @Kevin-Mattheus-Moerman
@fedorov, @ObedDzik, @zapaishchykova, @cnmy-ro, @thewtex, @lassoan, @antoinevacavant, @deepakri201,  would you be interested in reviewing this submission for [JOSS (the Journal of Open Source Software](https://joss.theoj.org)? The work is titled "_Slicer-Liver: A 3D Slicer Extension for Liver Surgery Planning_".

JOSS reviews take place on GitHub and focus on [the software :floppy_disk: ](https://github.com/ALive-research/Slicer-Liver) as well as [a short paper :scroll: ](https://raw.githubusercontent.com/openjournals/joss-papers/joss.07460/joss.07460/10.21105.joss.07460.pdf). We typically aim for a 2-3 week review but we can be more flexible if needed. You can find more about our review process in our [review guidelines](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html).  Here is an example of an ongoing review if you want to see what it is like: https://github.com/openjournals/joss-reviews/issues/7564.

If you are interested you can let me know by responding here.

Thanks!
