### 2025-02-06 #7460 @antoinevacavant
Hi,

I am collaborating with the team authoring this paper, I prefer to deny.

Best regards,
Antoine

On Thu, Feb 6, 2025 at 11:17 AM Andras Lasso ***@***.***>
wrote:

> I am a collaborator in the project, so I would not be an independent
> reviewer.
>
