### 2025-02-21 ALive-research/Slicer-Liver#274 @zapaishchykova
MarkupSlidingContour explanation in readme
Hi, thanks for the great work.  It would be great if you could add a bit more details in Readme under:
"In the 3D View, use the MarkupSlidingContour or MarkupDistanceContour tool to adjust the contour around the liver, or draw a closed curve with the MarkupClosedCurve tool." - a gif/video/mode details would help

Related to #7798
