### 2025-02-21 #7798 @zapaishchykova
Hi! Wonderful extension, I have a couple of minor suggections:
- “In the 3D View, use the MarkupSlidingContour or MarkupDistanceContour tool to adjust the contour around the liver, or draw a closed curve with the MarkupClosedCurve tool.” - Could authors please add more details on how to do that? https://github.com/ALive-research/Slicer-Liver/issues/274
- in statement of need, it would be helpful if authors specified who the target audience is and and its relation to other works in the field
- Please consider adding how this software compares to other commonly-used packages?
- Please add to the readme: Community guidelines: guidelines for third parties wishing to 1. Contribute to the software 2. Report issues or problems with the software 3. Seek support
