### 2025-02-22 ALive-research/Slicer-Liver#274 @RafaelPalomar
Thank you @zapaishchykova for the feedback. Here are some thoughts to improve the documentation:

-  Perhaps using _Markups-related_ terms is a little too technical in that part of the README. The user interface relates to _Contour Initialization_.
- There is an image with the 3 modes of initialization (a little bit below). It would be great if the image could be a video or gif of the 3 initialization modes in action, and if that could be placed closer to the text (or referenced).
- As a footnote/reference the concrete Markups implementing these could be mentioned, as this information could be useful for developers who would like to use our markups.

@RuoyanMeng, do you have capacity to take on these modifications?
