### 2025-03-01 74c1070
improve README and Joss paper
