+++ Paper/paper.md
@@ -43,11 +43,11 @@ This article introduces Slicer-Liver, an extension for the 3D Slicer image compu
 
 # Statement of need
 
-Liver cancer, both primary and metastatic (e.g., from colorectal cancer), is a global health concern with increasing incidence rates [@Siegel:2023] and surgical resection remains the most potentially curative treatment for some of these patients [@Petrowsky2020]. The evolution of computer-assisted surgical systems over the past two decades has significantly improved tumor localization and surgeons confidence during surgery [@Hansen:2014; @Lamata:2010]. Despite these advances, systematic use of computer-assisted systems for planning liver resections remains a challenge.
+Liver cancer, both primary and metastatic (e.g., from colorectal cancer), is a global health concern with increasing incidence rates [@Siegel:2023] and surgical resection remains the most potentially curative treatment for some of these patients [@Petrowsky2020]. The evolution of computer-assisted surgical systems over the past two decades has significantly improved tumor localization and surgeons confidence during surgery [@Hansen:2014; @Lamata:2010]. Despite these advances, systematic use of computer-assisted systems for planning liver resections remains a challenge. 
 
-Thanks to the latest advances in artificial intelligence, generation of 3D patient-specific models for surgical planning is increasingly becoming a reality in the clinical routine. However, surgical planning with the use of these models remains a complex and  manual process. Planning of surgery is particularly important for complex cases (e.g., those presenting multiple tumors or those where the location and size of the tumor poses a challenge for the surgery practice). Furthermore, precise surgical planning should not only account for the liver geometry, but also for the blood supply to various liver regions (segments) [@Warmann:2016; @Bismuth:2013]. Furthermore, visualization of 3D liver models and resections is a difficult task, where occlusions can prevent the effective understanding of the surgical plan. In addition, there is no broad consensus on the definition of a good resection plan, which is partly due to the lack of formal methods to specify and communicate resection plans, and partly due to the different surgery cultures and practices in different hospitals.
+Recent innovations in artificial intelligence have brought 3D patient-specific models closer to routine clinical use. However, surgical planning with the use of these models remains a complex and  manual process. Planning of surgery is particularly important for complex cases (e.g., those presenting multiple tumors or those where the location and size of the tumor poses a challenge for the surgery practice). Furthermore, precise surgical planning should not only account for the liver geometry, but also for the blood supply to various liver regions (segments) [@Warmann:2016; @Bismuth:2013]. Furthermore, visualization of 3D liver models and resections is a difficult task, where occlusions can prevent the effective understanding of the surgical plan. In addition, there is no broad consensus on the definition of a good resection plan, which is partly due to the lack of formal methods to specify and communicate resection plans, and partly due to the different surgery cultures and practices in different hospitals.
 
-All these challenges support the advancing on the integration of computer-aided systems for surgical planning into the clinical reality. In response to these challenges, Slicer-Liver provides functionality for modeling of resections and vascular territories, visualization of virtual resections in lower dimensions and volumetry analysis. In addition, Slicer-Liver serves as a research sandbox for integration of new liver therapy planning methods (e.g., liver ablations) and can be extended to planning of therapies in organs other than the liver.
[…]
