### 2025-03-01 ALive-research/Slicer-Liver#274 @RuoyanMeng
Hi, @zapaishchykova, I have created a pull request to clarify the corresponding context. I added a new sub-section in README to explain the difference and usage of contour initialization methods, including three gifs to showcase how those method works. Hope this helps 😃.
