### 2025-03-01 #7798 @RuoyanMeng
> Hi! Wonderful extension, I have a couple of minor suggestions:
>
> * “In the 3D View, use the MarkupSlidingContour or MarkupDistanceContour tool to adjust the contour around the liver, or draw a closed curve with the MarkupClosedCurve tool.” - Could authors please add more details on how to do that? [https://github.com/ALive-research/Slicer-Liver/issues/274   ](https://github.com/ALive-research/Slicer-Liver/issues/274%E2%80%A8)
> * in statement of need, it would be helpful if authors specified who the target audience is and and its relation to other works in the field
> * Please consider adding how this software compares to other commonly-used packages?
> * Please add to the readme: Community guidelines: guidelines for third parties wishing to 1. Contribute to the software 2. Report issues or problems with the software 3. Seek support

Hi @zapaishchykova,

Thank you for taking the time to review our work!
In response to your first two comments, I have created a pull request to clarify the corresponding context:
-I added a new subsection in the README to explain the differences and usage of contour initialization methods, including three GIFs to demonstrate how each method works.
-In JOSS paper.md, I specified our target audience at the end of the Statement of Need section.

Regarding the community guidelines, we do have a **Developer/Contributing** section in the README to guide those interested in contributing code, suggestions, or reporting bugs.

As for comparing our software with other commonly used packages, currently, no open-source solutions for liver surgery planning match the functionality offered by Slicer-Liver. While commercial tools like [MeVis](https://www.mevis.de/loesungen/mevis-liver-suite) and other research efforts ([1](https://doi.org/10.1016/j.cag.2024.01.006), [2](https://doi.org/10.1016/j.cmpb.2018.01.024)) focus on liver surgery planning, their lack of accessibility makes direct comparison challenging.

Let me know if you have any further suggestions!
