### 2025-03-02 #7798 @deepakri201
Hi,

This is a great extension -- I just have some minor comments:

- It might be useful to have a dialog box popup when a step is complete (after clicking Compute Distance Map for instance), or printing out a statement in the python console.
- In the Resectogram section in the readme, the following is stated: "Vascular Segments Volume: Use the dropdown menu to select the vascular segments volume. The available options depend on the pre-loaded volumes in the software." Would this be useful to include as sample data?
- I appreciate that a video tutorial is included. I noticed it's  using an older version of the extension. Could it be modified to include the latest changes to your extension?
- Automated tests from Reload+Test return errors
