### 2025-03-16 ba54ac0
improve README and Joss paper (#275)
