### 2025-03-16 60eb660
ENH: Add pop-up messages and cursor shape change to indicate the computing status; fix the unit-tests error message.
