+++ Liver/Liver.py
@@ -414,6 +414,9 @@ class LiverWidget(ScriptedLoadableModuleWidget):
     activeDistanceContour.SetNthControlPointPosition(1, tuple(point))
 
   def onMarkupsResectionCheckBoxChecked(self, checkbox):
+
+    qt.QApplication.setOverrideCursor(qt.Qt.WaitCursor)
+
     activeMarkupClosedCurveNode = self.resectionsWidget.MarkupClosedCurveNodeComboBox.currentNode()
     activeResectionNode = self.resectionsWidget.ResectionNodeComboBox.currentNode()
     liverNode = activeResectionNode.GetTargetOrganModelNode()
@@ -425,6 +428,8 @@ class LiverWidget(ScriptedLoadableModuleWidget):
                                                                                           liverNode))
       activeMarkupClosedCurveNode.AddObserver(slicer.vtkMRMLMarkupsNode.PointModifiedEvent,
                                               self.onDistanceContourStartInteraction)
+      qt.QApplication.restoreOverrideCursor()
+      qt.QMessageBox.information(None, "Information", "Resection surface initialization finished.")
 
   def onDistanceMapParameterChanged(self):
     """
@@ -737,13 +742,15 @@ class LiverWidget(ScriptedLoadableModuleWidget):
     downSamplingRate = self.distanceMapsWidget.DownsamplingRateSpinBox.value
     self.logic.computeDistanceMaps(tumorLabelmapVolumeNode, parenchymaLabelmapVolumeNode, hepaticLabelmapVolumeNode, portalLabelmapVolumeNode, outputVolumeNode, downSamplingRate)
 
+    # slicer.app.resumeRender()
     slicer.mrmlScene.RemoveNode(tumorLabelmapVolumeNode)
     slicer.mrmlScene.RemoveNode(parenchymaLabelmapVolumeNode)
     slicer.mrmlScene.RemoveNode(hepaticLabelmapVolumeNode)
     slicer.mrmlScene.RemoveNode(portalLabelmapVolumeNode)
 
-    slicer.app.resumeRender()
+    #slicer.app.resumeRender()
     qt.QApplication.restoreOverrideCursor()
+    qt.QMessageBox.information(None, "Information", "Distance maps computed.")
     slicer.util.showStatusMessage('')
 
   def onUncertaintyMaginComboBoxChanged(self):
