### 2025-03-16 653f3a1
ENH: Add pop-up messages and cursor shape change to indicate the computing status; fix the unit-tests error message.
