### 2025-03-18 #7798 @RuoyanMeng
> Hi,
>
> This is a great extension -- I just have some minor comments:
>
> * It might be useful to have a dialog box popup when a step is complete (after clicking Compute Distance Map for instance), or printing out a statement in the python console.
> * In the Resectogram section in the readme, the following is stated: "Vascular Segments Volume: Use the dropdown menu to select the vascular segments volume. The available options depend on the pre-loaded volumes in the software." Would this be useful to include as sample data?
> * I appreciate that a video tutorial is included. I noticed it's  using an older version of the extension. Could it be modified to include the latest changes to your extension?
> * Automated tests from Reload+Test return errors

Hi @deepakri201,

Apologies for the delayed response, and thank you for taking the time to review our work and share your feedback.

I've created a [pull request](https://github.com/ALive-research/Slicer-Liver/pull/277) addressing your first and last comments:
- **Indicating computation status**: Previously, we changed the cursor from ***arrow*** to ***waiting*** during computation (e.g., distance map processing) and reverted it once completed. However, as you suggested, this was not explicit enough. To improve clarity, we've added a popup dialog in several similar scenarios to inform users when the computation is complete.
- **Auto-tests error**: This issue has now been fixed.

Regarding the sample data for liver segments, it is hosted in a [separate repository](https://github.com/ALive-research/ALiveResearchTestingData), to which I don’t have direct edit access. I have shared the sample data with co-author @RafaelPalomar, and once it’s uploaded, I will update the Slicer-Liver repository accordingly within the same [pull request](https://github.com/ALive-research/Slicer-Liver/pull/277).

Additionally, co-author @dalbenzioG is working on updating the tutorial video to reflect the latest changes in our extension.

Please let me know if you have any additional feedback!
