### 2025-03-21 b526144
ENH: Add VascularSegments as sample data
