### 2025-03-21 fb2ec88
ENH: Add VascularSegments as sample data
Co-authored-by: Rafael Palomar <rafael.palomar@ous-research.no>
