+++ LiverVolumetry/LiverVolumetry.py
@@ -422,8 +422,14 @@ class LiverVolumetryLogic(ScriptedLoadableModuleLogic):
         segStatLogic.computeStatistics()
         stats = segStatLogic.getStatistics()
         for segmentId in stats["SegmentIDs"]:
-          voxel_count = stats[segmentId,"LabelmapSegmentStatisticsPlugin.voxel_count"]
-          volume_cm3 = stats[segmentId,"LabelmapSegmentStatisticsPlugin.volume_cm3"]
+          voxel_count = 0
+          volume_cm3 = 0
+          if stats[segmentId,"LabelmapSegmentStatisticsPlugin.voxel_count"]:
+            voxel_count = stats[segmentId,"LabelmapSegmentStatisticsPlugin.voxel_count"]
+            volume_cm3 = stats[segmentId,"LabelmapSegmentStatisticsPlugin.volume_cm3"]
+          elif stats[segmentId,"ScalarVolumeSegmentStatisticsPlugin.voxel_count"]:
+            voxel_count = stats[segmentId,"ScalarVolumeSegmentStatisticsPlugin.voxel_count"]
+            volume_cm3 = stats[segmentId,"ScalarVolumeSegmentStatisticsPlugin.volume_cm3"]
           segmentName = segmentationNode.GetSegmentation().GetSegment(segmentId).GetName()
           statistics[segmentId] = [segmentName, voxel_count, volume_cm3]
           self.scl.VolumetryTable(segmentName, targetSegmentVolume, voxel_count, volume_cm3,outputTable)
@@ -441,11 +447,18 @@ class LiverVolumetryLogic(ScriptedLoadableModuleLogic):
     else:
       self.scl.ComputeAdvancedPlanningVolumetry(segmentsVolumeNode, outputTable, ROIMarkersList, resectionNodes, targetSegmentVolume)
 
+  def generateSegments(self, resectionNodes, ROIMarkersList, segmentsVolumeNode):
+    generatedSegmentsNode = slicer.mrmlScene.AddNewNodeByClass("vtkMRMLLabelMapVolumeNode")
+    generatedSegmentsNode.CreateDefaultDisplayNodes()
 
+    self.scl.GenerateSegmentsLabelMap(segmentsVolumeNode, generatedSegmentsNode, resectionNodes, ROIMarkersList)
 
+    seg = slicer.mrmlScene.AddNewNodeByClass("vtkMRMLSegmentationNode")
+    slicer.modules.segmentations.logic().ImportLabelmapToSegmentationNode(generatedSegmentsNode, seg)
 
+    ##set segments label
+    seg.GetSegmentation().GetNthSegment(0).SetName("Remnant")
+    for i in range(ROIMarkersList.GetNumberOfControlPoints()):
+      seg.GetSegmentation().GetNthSegment(i+1).SetName(ROIMarkersList.GetNthFiducialLabel(i))
 
-
-
-
-
+    slicer.mrmlScene.RemoveNode(generatedSegmentsNode)
