### 2025-03-21 8e1a6c4
ENH: Add VascularSegments as sample data
Co-authored-by: Rafael Palomar <rafael.palomar@ous-research.no>
