### 2025-04-25 #7798 @RuoyanMeng
Hi @Kevin-Mattheus-Moerman,
We created a [GitHub release](https://github.com/ALive-research/Slicer-Liver/releases/tag/v1.1.0) and archived it via Zenodo, doi: [10.5281/zenodo.15279062](https://doi.org/10.5281/zenodo.15279062).
One small thing I found is that in Zenodo, the auto-generated affiliation index is different from what we have in the JOSS paper. Does this matter? Should we adjust the JOSS paper to match the Zenodo one? Thanks and looking forward to our next step!
