### 2025-04-25 #7798 @Kevin-Mattheus-Moerman
@RuoyanMeng I checked the paper, could you have a look at the below:
- [ ] Check spelling for `resetions`, should this perhaps be `resections`?
