### 2025-04-25 47736e0
DOC: Minor fix of paper
