+++ Paper/paper.md
@@ -39,28 +39,28 @@ bibliography: bibliography.bib
 
 # Summary
 
-This article introduces Slicer-Liver, an extension for the 3D Slicer image computing platform [@Kikinis:2013] ([https://slicer.org](https://slicer.org)). The software aims to address challenges in liver surgery planning by leveraging geometric modeling and visualization techniques. More specifically, Slicer-Liver enables integral liver surgery plans including virtual resections, vascular territories, as well as volumetry analysis. Progress and preliminary results obtained with Slicer-Liver show the potential of the software to be used for improving the state-of-the-art approaches in definition of virtual resections, visualization of resection plans, as well as classification of the liver in segments. These contributions hold promise on enhancing liver surgery planning and potentially improving liver surgery outcomes. Slicer-Liver is intended for research use within the fields of biomedicine and computer science.
+This article introduces Slicer-Liver, an extension for the 3D Slicer image computing platform [@Kikinis:2013] ([https://slicer.org](https://slicer.org)). The software aims to address challenges in liver surgery planning by leveraging geometric modeling and visualization techniques. More specifically, Slicer-Liver enables integral liver surgery plans, including virtual resections, vascular territories, as well as volumetry analysis. Progress and preliminary results obtained with Slicer-Liver show the potential of the software to be used for improving the state-of-the-art approaches in the definition of virtual resections, visualization of resection plans, as well as classification of the liver segments. These contributions hold promise on enhancing liver surgery planning and potentially improving liver surgery outcomes. Slicer-Liver is intended for research use within the fields of biomedicine and computer science.
 
 # Statement of need
 
-Liver cancer, both primary and metastatic (e.g., from colorectal cancer), is a global health concern with increasing incidence rates [@Siegel:2023] and surgical resection remains the most potentially curative treatment for some of these patients [@Petrowsky2020]. The evolution of computer-assisted surgical systems over the past two decades has significantly improved tumor localization and surgeons confidence during surgery [@Hansen:2014; @Lamata:2010]. Despite these advances, systematic use of computer-assisted systems for planning liver resections remains a challenge. 
+Liver cancer, both primary and metastatic (e.g., from colorectal cancer), is a global health concern with increasing incidence rates [@Siegel:2023] and surgical resection remains the most potentially curative treatment for some of these patients [@Petrowsky2020]. The evolution of computer-assisted surgical systems over the past two decades has significantly improved tumor localization and surgeons' confidence during surgery [@Hansen:2014; @Lamata:2010]. Despite these advances, systematic use of computer-assisted systems for planning liver resections remains a challenge. 
 
-Recent innovations in artificial intelligence have brought 3D patient-specific models closer to routine clinical use. However, surgical planning with the use of these models remains a complex and  manual process. Planning of surgery is particularly important for complex cases (e.g., those presenting multiple tumors or those where the location and size of the tumor poses a challenge for the surgery practice). Furthermore, precise surgical planning should not only account for the liver geometry, but also for the blood supply to various liver regions (segments) [@Warmann:2016; @Bismuth:2013]. Furthermore, visualization of 3D liver models and resections is a difficult task, where occlusions can prevent the effective understanding of the surgical plan. In addition, there is no broad consensus on the definition of a good resection plan, which is partly due to the lack of formal methods to specify and communicate resection plans, and partly due to the different surgery cultures and practices in different hospitals.
[…]
