### 2025-04-27 #7798 @editorialbot
:wave: @openjournals/bcm-eics, this paper is ready to be accepted and published.
